Add flow for calculating the number of minutes available for fargate instance based on user's balance

+++ src/modules/user/infrastructure/persistence/user.schema.ts
@@ -26,7 +26,7 @@ export const UserSchema = new EntitySchema<User>({
     },
     balance: {
       name: 'balance',
-      type: 'int',
+      type: 'float',
       nullable: true,
     },
   },
